+++ src/component/admin/language/en-GB/com_cmsmigrator.ini
@@ -2,13 +2,18 @@ COM_CMSMIGRATOR = "Migrate To Joomla"
 COM_CMSMIGRATOR_DESCRIPTION = "This component migrates websites from other platforms to Joomla."
 COM_CMSMIGRATOR_MENU_BACKEND = "Migrate to Joomla"
 COM_CMSMIGRATOR_IMPORT_FILE_ERROR="Error uploading import file."
-COM_CMSMIGRATOR_IMPORT_SUCCESS="File uploaded successfully. Processing..."
+COM_CMSMIGRATOR_IMPORT_SUCCESS="Import successful."
 COM_CMSMIGRATOR_INVALID_JSON_FILE="The uploaded file is not a valid JSON file."
-COM_CMSMIGRATOR_NO_PLUGIN_FOUND="No migration plugin found for '%s'."
-COM_CMSMIGRATOR_INVALID_JSON_FORMAT_FROM_PLUGIN="The migration plugin returned an invalid JSON format."
-COM_CMSMIGRATOR_INVALID_ITEMLIST_FORMAT="The imported data is not a valid ItemList."
+COM_CMSMIGRATOR_NO_PLUGIN_FOUND="No suitable migration plugin found for %s."
+COM_CMSMIGRATOR_INVALID_JSON_FORMAT_FROM_PLUGIN="The plugin returned an invalid JSON format."
+COM_CMSMIGRATOR_INVALID_ITEMLIST_FORMAT="The provided data is not a valid ItemList format."
 COM_CMSMIGRATOR_UNCATEGORISED_NOT_FOUND="The default 'Uncategorised' category was not found."
-COM_CMSMIGRATOR_ARTICLE_IMPORTED_SUCCESS="Successfully imported article: %s"
+COM_CMSMIGRATOR_ARTICLE_IMPORTED_SUCCESS="Article '%s' imported successfully."
 COM_CMSMIGRATOR_FORM_NOT_FOUND="The import form could not be loaded."
 COM_CMSMIGRATOR_MANAGER_CPANEL="CMS Migration Tool"
 COM_CMSMIGRATOR_IMPORT_BUTTON="Import"
+COM_CMSMIGRATOR_UNABLE_TO_LOAD_ARTICLE_TABLE="Could not load article table."
+COM_CMSMIGRATOR_UNABLE_TO_LOAD_CATEGORY_TABLE="Could not load category table."
+
+COM_CMSMIGRATOR_JSON_SAVED="Imported JSON data saved to %s"
+COM_CMSMIGRATOR_JSON_SAVE_FAILED="Failed to save imported JSON data: %s"
