+++ src/component/admin/src/Model/ProcessorModel.php
@@ -1,0 +1,234 @@
+<?php
+
+namespace Binary\Component\CmsMigrator\Administrator\Model;
+
+\defined('_JEXEC') or die;
+
+use Joomla\CMS\Date\Date;
+use Joomla\CMS\Factory;
+use Joomla\CMS\Language\Text;
+use Joomla\CMS\MVC\Model\BaseDatabaseModel;
+use Joomla\CMS\Table\Table;
+use Joomla\CMS\Filesystem\Folder;
+use Joomla\CMS\Filesystem\File;
+use Joomla\CMS\Http\Http;
+
+class ProcessorModel extends BaseDatabaseModel
+{
+    protected $sourceUrl;
+    protected $http;
+
+    public function __construct(string $sourceUrl, Http $http, array $config = [])
+    {
+        $this->sourceUrl = rtrim($sourceUrl, '/');
+        $this->http = $http;
+
+        parent::__construct($config);
+    }
+
+    public function process(array $data)
+    {
+        if (!isset($data['@type']) || $data['@type'] !== 'ItemList' || !isset($data['itemListElement'])) {
+            $this->setError(Text::_('COM_CMSMIGRATOR_INVALID_ITEMLIST_FORMAT'));
+            return false;
+        }
+
+        foreach ($data['itemListElement'] as $element) {
+            if ($element['item']['@type'] === 'Article') {
+                $this->importArticle($element['item']);
+            }
+        }
+
+        return true;
+    }
+
+    protected function importArticle(array $articleData)
+    {
+        $articleTable = Table::getInstance('Content', 'Joomla\\Component\\Content\\Administrator\\Table\\ContentTable');
+
+        if ($articleTable === false) {
+            $this->setError(Text::_('COM_CMSMIGRATOR_UNABLE_TO_LOAD_ARTICLE_TABLE'));
+            return false;
+        }
+
+        // Get the first category name, or use an empty string as a fallback for 'uncategorised'
+        $categoryName = $articleData['articleSection'][0] ?? '';
+        $authorName = $articleData['author']['name'] ?? '';
+        $articleBody = $articleData['articleBody'] ?? '';
+
+        // Process images if a source URL is provided
+        if (!empty($this->sourceUrl)) {
+            $articleBody = $this->processImages($articleBody);
+        }
+
+        $data = [];
+        $data['title'] = $articleData['headline'] ?? 'Untitled Article';
+        $data['articletext'] = $articleBody;
+        $data['created'] = isset($articleData['datePublished']) ? (new Date($articleData['datePublished']))->toSql() : Factory::getDate()->toSql();
+        $data['created_by'] = $this->getAuthorId($authorName);
+        $data['state'] = 1;
+        $data['language'] = '*';
+        $data['catid'] = $this->getCategoryId($categoryName);
+
+
+        if (!$articleTable->bind($data)) {
+            $this->setError($articleTable->getError());
+            return false;
+        }
+
+        if (!$articleTable->check()) {
+            $this->setError($articleTable->getError());
+            return false;
+        }
+
+        if (!$articleTable->store()) {
+            $this->setError($articleTable->getError());
+            return false;
+        }
+
+        Factory::getApplication()->enqueueMessage(Text::sprintf('COM_CMSMIGRATOR_ARTICLE_IMPORTED_SUCCESS', $data['title']));
+
+        return true;
+    }
+
+    protected function processImages(string $text): string
+    {
+        if (empty($text)) {
+            return '';
+        }
+
+        preg_match_all('/<img[^>]+src="([^">]+)"/i', $text, $matches);
+
+        if (empty($matches[1])) {
+            return $text;
+        }
+
+        $destinationFolder = JPATH_SITE . '/images/imported_media';
+        Folder::create($destinationFolder);
+
+        foreach ($matches[1] as $originalSrc) {
+            $imageUrl = $originalSrc;
+            // Handle relative URLs
+            if (strpos($imageUrl, 'http') !== 0) {
+                if ($imageUrl[0] !== '/') {
+                    $imageUrl = '/' . $imageUrl;
+                }
+                $imageUrl = $this->sourceUrl . $imageUrl;
+            }
+
+            try {
+                $response = $this->http->get($imageUrl);
+
+                if ($response->code !== 200) {
+                    continue;
+                }
+
+                $filename = basename(parse_url($imageUrl, PHP_URL_PATH));
+                $destinationPath = $destinationFolder . '/' . $filename;
+
+                if (File::write($destinationPath, $response->body)) {
+                    $newSrc = 'images/imported_media/' . $filename;
+                    $text = str_replace($originalSrc, $newSrc, $text);
+                }
+            } catch (\Exception $e) {
+                // Could log the error, but for now we just skip the image
+                continue;
+            }
+        }
+
+        return $text;
+    }
+
+    protected function getAuthorId(string $authorName): int
+    {
+        if (empty($authorName)) {
+            return Factory::getUser()->id;
+        }
+
+        $db = $this->getDbo();
+        $query = $db->getQuery(true)
+            ->select($db->quoteName('id'))
+            ->from($db->quoteName('#__users'))
+            ->where($db->quoteName('username') . ' = ' . $db->quote($authorName) . ' OR ' . $db->quoteName('name') . ' = ' . $db->quote($authorName));
+
+        $authorId = $db->setQuery($query)->loadResult();
+
+        if ($authorId) {
+            return (int) $authorId;
+        }
+
+        // Fallback to current user if no matching author is found
+        return Factory::getUser()->id;
+    }
+
+    protected function getCategoryId(string $categoryName): int
+    {
+        // Use default 'Uncategorised' if no name is provided
+        if (empty($categoryName)) {
+            $uncategorisedTable = Table::getInstance('Category', 'Joomla\\Component\\Categories\\Administrator\\Table\\CategoryTable');
+
+            if ($uncategorisedTable === false) {
+                // Should not happen on a standard Joomla install
+                return 1;
+            }
+
+            if ($uncategorisedTable->load(['alias' => 'uncategorised', 'extension' => 'com_content'])) {
+                return (int) $uncategorisedTable->id;
+            }
+            // Should not happen on a standard Joomla install
+            return 1;
+        }
+
+        $db = $this->getDbo();
+        $query = $db->getQuery(true)
+            ->select($db->quoteName('id'))
+            ->from($db->quoteName('#__categories'))
+            ->where($db->quoteName('title') . ' = ' . $db->quote($categoryName))
+            ->where($db->quoteName('extension') . ' = ' . $db->quote('com_content'));
+        
+        $categoryId = $db->setQuery($query)->loadResult();
+
+        if ($categoryId) {
+            return (int) $categoryId;
+        }
+
+        // Category not found, create it as a top-level category
+        $categoryTable = Table::getInstance('Category', 'Joomla\\Component\\Categories\\Administrator\\Table\\CategoryTable');
+
+        if ($categoryTable === false) {
+            $this->setError(Text::_('COM_CMSMIGRATOR_UNABLE_TO_LOAD_CATEGORY_TABLE'));
+            // Fallback to Uncategorised on failure
+            return $this->getCategoryId('');
+        }
+        
+        $categoryData = [
+            'title' => $categoryName,
+            'alias' => \Joomla\CMS\Filter\OutputFilter::stringURLSafe($categoryName),
+            'extension' => 'com_content',
+            'published' => 1,
+            'language' => '*',
+            'access' => 1, // Public
+            'parent_id' => 1, // Root
+            'created_user_id' => Factory::getUser()->id,
+        ];
+
+        if (!$categoryTable->bind($categoryData)) {
+            $this->setError($categoryTable->getError());
+            return $this->getCategoryId('');
+        }
+
+        if (!$categoryTable->check()) {
+            $this->setError($categoryTable->getError());
+            return $this->getCategoryId('');
+        }
+
+        if (!$categoryTable->store()) {
+            $this->setError($categoryTable->getError());
+            return $this->getCategoryId('');
+        }
+        
+        $categoryTable->rebuildPath($categoryTable->id);
+
+        return (int) $categoryTable->id;
+    }
+}
